+++ ArrayDefinitions.h
@@ -9,73 +9,70 @@
 #define ARRAY_DEFINITIONS_H
 
 
-template <typename T, unsigned int max_size_>
+template <typename T, size_t max_size_>
 Array<T,max_size_>::Array()
 {
   clear();
 }
 
-template <typename T, unsigned int max_size_>
+template <typename T, size_t max_size_>
 Array<T,max_size_>::Array(const T &value)
 {
   fill(value);
 }
 
-template <typename T, unsigned int max_size_>
+template <typename T, size_t max_size_>
 Array<T, max_size_>::Array(const T (&values)[max_size_])
 {
   fill(values);
 }
 
-template <typename T, unsigned int max_size_>
-T& Array<T, max_size_>::operator[](const unsigned int i)
+template <typename T, size_t max_size_>
+T& Array<T, max_size_>::operator[](const size_t i)
 {
   return values_[i];
 }
 
-template <typename T, unsigned int max_size_>
-T& Array<T, max_size_>::at(const unsigned int i)
+template <typename T, size_t max_size_>
+T& Array<T, max_size_>::at(const size_t i)
 {
-  // assert((i>=0) && (i<size_));
   return values_[i];
 }
 
-template <typename T, unsigned int max_size_>
+template <typename T, size_t max_size_>
 T& Array<T, max_size_>::front()
 {
-  // assert(!empty());
   return values_[0];
 }
 
-template <typename T, unsigned int max_size_>
+template <typename T, size_t max_size_>
 T& Array<T, max_size_>::back()
 {
-  // assert(!empty());
   return values_[size_-1];
 }
 
-template <typename T, unsigned int max_size_>
+template <typename T, size_t max_size_>
 void Array<T, max_size_>::clear()
 {
   size_ = 0;
 }
 
-template <typename T, unsigned int max_size_>
+template <typename T, size_t max_size_>
 void Array<T, max_size_>::fill(const T &value)
 {
   assign(max_size_,value);
 }
 
-template <typename T, unsigned int max_size_>
+template <typename T, size_t max_size_>
 void Array<T, max_size_>::fill(const T (&values)[max_size_])
 {
   assign(max_size_,values);
 }
 
-template <typename T, unsigned int max_size_>
-void Array<T, max_size_>::assign(const unsigned int n, const T &value)
+template <typename T, size_t max_size_>
+void Array<T, max_size_>::assign(const size_t n, const T &value)
 {
-  unsigned int assign_size = n;
+  size_t assign_size = n;
   if ((n > size_) && (n <= max_size_))
   {
     size_ = n;
@@ -85,16 +82,16 @@ void Array<T, max_size_>::assign(const unsigned int n, const T &value)
     size_ = max_size_;
     assign_size = max_size_;
   }
-  for (unsigned int i=0; i<assign_size; i++)
+  for (size_t i=0; i<assign_size; i++)
   {
     values_[i] = value;
   }
 }
 
-template <typename T, unsigned int max_size_>
-void Array<T, max_size_>::assign(const unsigned int n, const T values[])
+template <typename T, size_t max_size_>
+void Array<T, max_size_>::assign(const size_t n, const T values[])
 {
-  unsigned int assign_size = n;
+  size_t assign_size = n;
   if ((n > size_) && (n <= max_size_))
   {
     size_ = n;
@@ -107,7 +104,7 @@ void Array<T, max_size_>::assign(const unsigned int n, const T values[])
   memcpy((void*) values_, (void*) values, assign_size*sizeof(T));
 }
 
-template <typename T, unsigned int max_size_>
+template <typename T, size_t max_size_>
 void Array<T, max_size_>::push_back(const T &value)
 {
   if (size_ < max_size_)
@@ -116,7 +113,7 @@ void Array<T, max_size_>::push_back(const T &value)
   }
 }
 
-template <typename T, unsigned int max_size_>
+template <typename T, size_t max_size_>
 void Array<T, max_size_>::pop_back()
 {
   if (size_ > 0)
@@ -125,25 +122,25 @@ void Array<T, max_size_>::pop_back()
   }
 }
 
-template <typename T, unsigned int max_size_>
-unsigned int Array<T, max_size_>::size()
+template <typename T, size_t max_size_>
+size_t Array<T, max_size_>::size()
 {
   return size_;
 }
 
-template <typename T, unsigned int max_size_>
-unsigned int Array<T, max_size_>::max_size()
+template <typename T, size_t max_size_>
+size_t Array<T, max_size_>::max_size()
 {
   return max_size_;
 }
 
-template <typename T, unsigned int max_size_>
+template <typename T, size_t max_size_>
 bool Array<T, max_size_>::empty()
 {
   return size_ == 0;
 }
 
-template <typename T, unsigned int max_size_>
+template <typename T, size_t max_size_>
 bool Array<T, max_size_>::full()
 {
   return size_ == max_size_;
